--- src/core/internal/viem/core.ts
@@ -1,1 +1,0 @@
-// TODO
